+++ src/Exceptions/InvalidJournalEntryValue.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Academe\LaravelJournal\Exceptions;
 
-class InvalidJournalEntryValue extends JournalException
+class InvalidJournalEntryValue extends JournalLogicException
 {
     public function __construct(string $message = 'Journal transaction entries must be a positive value.')
     {