+++ src/Exceptions/InvalidJournalMethod.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Academe\LaravelJournal\Exceptions;
 
-class InvalidJournalMethod extends JournalException
+class InvalidJournalMethod extends JournalLogicException
 {
     public function __construct(string $message = 'Journal methods must be credit or debit.')
     {