+++ src/Exceptions/InvalidTags.php
@@ -4,4 +4,4 @@ declare(strict_types=1);
 
 namespace Academe\LaravelJournal\Exceptions;
 
-class InvalidTags extends JournalException {}
+class InvalidTags extends JournalLogicException {}