+++ src/Exceptions/JournalAlreadyExists.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Academe\LaravelJournal\Exceptions;
 
-class JournalAlreadyExists extends JournalException
+class JournalAlreadyExists extends JournalRuntimeException
 {
     public function __construct(string $message = 'Journal already exists for this model.')
     {